Sort Options and Outputs in API.txt
Python does not guarantee dict order. Our reliance on it in `makeapi`
is technically incorrect, even if it doesn't matter yet in any of our
developers' environments.
This patch sorts the options by name instead of relying on dict order.

As an added benefit, future API.txt diffs should be be neater.